\name{bbCrosstab}
\alias{bbCrosstab}
\title{Create pairwise crosstabulation}
\description{
  Create the pairwise crosstabulation of two or more factors and store
  it as an object of class \code{"bbCrosstab"} which consists of a
  list of sparse, blocked matrices.
}
\usage{
bbCrosstab(flist)
}
\arguments{
  \item{flist}{A list of two or more factors.  All factors must have the
    same length.}
}
%\details{}
\value{
  An object of class \code{"bbCrosstab"}.
}
%\references{}
%\author{}
%\note{}
\seealso{\code{\link{bbCrosstab-class}}}
%\examples{}
\keyword{manip}
\keyword{algebra}
