\name{Complex-methods}
\title{Unary Operations on Complex Numbers}
%
\docType{methods}
\keyword{complex}
\keyword{methods}
%
\alias{Complex}
\alias{Complex-methods}
%
\alias{Summary,denseMatrix-method}
\alias{Summary,diagonalMatrix-method}
\alias{Summary,indMatrix-method}
\alias{Summary,sparseMatrix-method}
\alias{Summary,sparseVector-method}
%
\description{TODO.}
\usage{
Complex(z)
}
\arguments{
\item{z}{.}
}
%% \value{}
%% \seealso{}
%% \examples{}
