\name{corFactor}
\docType{genericFunction}
\alias{corFactor}
\title{Extract the correlation factor}
\description{
  A generic function to extract the factor of the correlation matrix of
  the object represented by \code{object}.
}
\usage{
corFactor(object, ...)
}
\arguments{
  \item{object}{an object that can be coerced to a correlation matrix factor}
  \item{\dots}{some methods for this generic may take additional,
    optional arguments}
}
%\details{}
\value{
  An object of class \code{corrfactor}.
}
%\references{}
\author{Douglas Bates \email{bates@stat.wisc.edu}
  and Saikat DebRoy \email{saikat@stat.wisc.edu}}
%\note{}
%\seealso{}
\examples{
}
\keyword{models}
