\name{dimnames-methods}
\title{Dimnames of a Matrix}
%
\docType{methods}
\keyword{array}
\keyword{attribute}
\keyword{methods}
%
\alias{dimnames}
\alias{dimnames-methods}
\alias{dimnames<-}
\alias{dimnames<--methods}
%
\alias{dimnames,Matrix-method}
\alias{dimnames,MatrixFactorization-method}
\alias{dimnames,sparseVector-method}
\alias{dimnames,symmetricMatrix-method}
%
\alias{dimnames<-,Matrix,NULL-method}
\alias{dimnames<-,Matrix,list-method}
\alias{dimnames<-,MatrixFactorization,NULL-method}
\alias{dimnames<-,MatrixFactorization,list-method}
\alias{dimnames<-,generalMatrix,NULL-method}
\alias{dimnames<-,generalMatrix,list-method}
\alias{dimnames<-,symmetricMatrix,NULL-method}
\alias{dimnames<-,symmetricMatrix,list-method}
%
\description{TODO.}
\usage{
\S4method{dimnames}{Matrix}(x)
\S4method{dimnames}{MatrixFactorization}(x)

\S4method{dimnames}{Matrix,list}(x) <- value
\S4method{dimnames}{MatrixFactorization,list}(x) <- value

\S4method{unname}{Matrix}(obj, force = FALSE)
\S4method{unname}{MatrixFactorization}(obj, force = FALSE)
}
\arguments{
\item{x}{.}
\item{value}{.}
\item{obj}{.}
\item{force}{.}
}
%% \value{}
%% \seealso{}
%% \examples{}
