% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{EM}
\alias{EM}
\title{estimate parameters of EM algorithm}
\usage{
EM(data, init.value)
}
\arguments{
\item{data}{a vector of data}

\item{init.value}{initial value for the vector p.}
}
\value{
estimated parameters of p
}
\description{
estimate parameters of EM algorithm
}
