\name{FastCoercions}
\title{}
%
\alias{FastCoercions}
%
\alias{.CR2RC}
\alias{.CR2T}
\alias{.M2diag}
\alias{.M2sym}
\alias{.M2tri}
\alias{.T2CR}
\alias{.dense2g}
\alias{.dense2kind}
\alias{.dense2m}
\alias{.dense2sparse}
\alias{.diag2kind}
\alias{.diag2m}
\alias{.diag2sparse}
\alias{.m2dense}
\alias{.m2sparse}
\alias{.sparse2dense}
\alias{.sparse2g}
\alias{.sparse2kind}
\alias{.sparse2m}
\alias{.tCR2RC}
%
\description{
}
\usage{
.CR2RC(from)
.CR2T(from)
.M2diag(from, check = TRUE)
.M2sym(from, ...)
.M2tri(from, ...)
.T2CR(from, Csparse = TRUE)
.dense2g(from, kind = ".")
.dense2kind(from, kind)
.dense2m(from, ndense = FALSE)
.dense2sparse(from, code)
.diag2kind(from, kind)
.diag2m(from)
.diag2sparse(from, code, uplo = "U", drop0 = TRUE)
.m2dense(from, code, uplo = "U", diag = "N")
.m2sparse(from, code, uplo = "U", diag = "N")
.sparse2dense(from, packed = FALSE)
.sparse2g(from)
.sparse2kind(from, kind, drop0 = FALSE)
.sparse2m(from)
.tCR2RC(from)
}
\arguments{
  \item{}{}
}
\value{
}
\details{
}
\seealso{
}
\keyword{internal}
