\name{forceDiagonal-methods}
\title{Get the Diagonal Part of a Matrix}
%
\docType{methods}
\keyword{array}
\keyword{methods}
%
\alias{forceDiagonal}
\alias{forceDiagonal-methods}
%
\alias{forceDiagonal,CsparseMatrix-method}
\alias{forceDiagonal,RsparseMatrix-method}
\alias{forceDiagonal,TsparseMatrix-method}
\alias{forceDiagonal,denseMatrix-method}
\alias{forceDiagonal,diagonalMatrix-method}
\alias{forceDiagonal,indMatrix-method}
\alias{forceDiagonal,matrix-method}
%
\description{TODO.}
\usage{
forceDiagonal(x, \dots)
}
\arguments{
\item{x}{a matrix or \code{\linkS4class{Matrix}}.}
\item{\dots}{optional arguments passed from or to other methods.}
}
%% \value{}
%% \seealso{}
%% \examples{}
