\name{forceTriangular-methods}
\title{Get the Upper or Lower Triangular Part of a Square Matrix}
%
\docType{methods}
\keyword{array}
\keyword{methods}
%
\alias{forceTriangular}
\alias{forceTriangular-methods}
%
\alias{forceTriangular,CsparseMatrix-method}
\alias{forceTriangular,RsparseMatrix-method}
\alias{forceTriangular,TsparseMatrix-method}
\alias{forceTriangular,denseMatrix-method}
\alias{forceTriangular,diagonalMatrix-method}
\alias{forceTriangular,indMatrix-method}
\alias{forceTriangular,matrix-method}
%
\description{TODO.}
\usage{
forceTriangular(x, \dots)
}
\arguments{
\item{x}{a square matrix or \code{\linkS4class{Matrix}}.}
\item{\dots}{optional arguments passed from or to other methods.}
}
%% \value{}
%% \seealso{}
%% \examples{}
