\name{head-methods}
\title{Head and Tail of a Matrix or Vector}
%
\docType{methods}
\keyword{array}
\keyword{methods}
%
\alias{head}
\alias{head-methods}
\alias{tail}
\alias{tail-methods}
%
\alias{head,Matrix-method}
\alias{head,sparseVector-method}
%
\alias{tail,Matrix-method}
\alias{tail,sparseVector-method}
%
\description{TODO.}
\usage{
head(x, \dots)
\S4method{head}{Matrix}(x, n = 6L, \dots)
\S4method{head}{sparseVector}(x, n = 6L, \dots)

tail(x, \dots)
\S4method{tail}{Matrix}(x, n = 6L, \dots)
\S4method{tail}{sparseVector}(x, n = 6L, \dots)
}
\arguments{
\item{x}{.}
\item{n}{.}
\item{\dots}{.}
}
%% \value{}
%% \seealso{}
%% \examples{}
