\name{initialize-methods}
\title{Initialization}
%
\docType{methods}
\keyword{methods}
%
\alias{initialize}
\alias{initialize-methods}
%
\alias{initialize,Matrix-method}
\alias{initialize,MatrixFactorization-method}
\alias{initialize,sparseVector-method}
%
\description{TODO.}
\usage{
\S4method{initialize}{Matrix}(.Object, \dots)
\S4method{initialize}{MatrixFactorization}(.Object, \dots)
\S4method{initialize}{sparseVector}(.Object, i, x, \dots)
}
\arguments{
\item{.Object}{.}
\item{i}{.}
\item{x}{.}
\item{\dots}{.}
}
%% \value{}
%% \seealso{}
%% \examples{}
