\name{Math-methods}
\title{Mathematical Functions}
%
\docType{methods}
\keyword{math}
\keyword{methods}
%
\alias{Math}
\alias{Math-methods}
\alias{Math2}
\alias{Math2-methods}
%
\alias{Math,denseMatrix-method}
\alias{Math,diagonalMatrix-method}
\alias{Math,indMatrix-method}
\alias{Math,sparseMatrix-method}
\alias{Math,sparseVector-method}
%
\alias{Math2,Matrix-method}
\alias{Math2,sparseVector-method}
%
\alias{log,denseMatrix-method}
\alias{log,diagonalMatrix-method}
\alias{log,indMatrix-method}
\alias{log,sparseMatrix-method}
\alias{log,sparseVector-method}
%
\alias{zapsmall,Matrix-method}
\alias{zapsmall,sparseVector-method}
%
\description{TODO.}
\usage{
Math(x)
Math2(x, digits)
}
\arguments{
\item{x}{.}
\item{digits}{.}
}
%% \value{}
%% \seealso{}
%% \examples{}
