\name{Math2-methods}
\title{Rounding of Numbers}
%
\docType{methods}
\keyword{math}
\keyword{methods}
%
\alias{Math2}
\alias{Math2-methods}
%
\alias{Math2,denseMatrix-method}
\alias{Math2,diagonalMatrix-method}
\alias{Math2,indMatrix-method}
\alias{Math2,sparseMatrix-method}
\alias{Math2,sparseVector-method}
%
\description{TODO.}
\usage{
Math2(x, digits)
}
\arguments{
\item{x}{.}
\item{digits}{.}
}
%% \value{}
%% \seealso{}
%% \examples{}
