\name{Matrix-notyet}
\title{Virtual Classes Not Yet Really Implemented and Used}
%
\docType{class}
\keyword{array}
\keyword{classes}
%
\alias{Matrix-notyet}
%
\alias{iMatrix-class}
\alias{zMatrix-class}
%
\description{
  \code{iMatrix} is the virtual class of all integer
  (S4) matrices.  It extends the \code{\linkS4class{Matrix}} class directly.

  \code{zMatrix} is the virtual class of all \code{\link{complex}}
  (S4) matrices.  It extends the \code{\linkS4class{Matrix}} class directly.
}
\examples{
showClass("iMatrix")
showClass("zMatrix")
}
