\name{matrixGets}
\docType{genericFunction}
\alias{matrix<-}
\alias{matrix<--methods}
\alias{matrix<-,pdMat,matrix-method}
\title{Assign the matrix}
\description{
  This generic function assigns the matrix to an object of a suitable
  class, such as the \code{pdMat} class.

  This function is deprecated.  Use a coerced assignment instead.
}
\usage{
matrix(object) <- value
}
\arguments{
  \item{object}{An object of a suitable class}
  \item{value}{A matrix of a suitable size and family.}
}
\value{
  An object of the same class as \code{object}.
}
%\references{}
%\author{Douglas Bates \email{bates@stat.wisc.edu}
%  and Saikat DebRoy \email{saikat@stat.wisc.edu}}
%\note{}
%\seealso{}
%\examples{}
\keyword{manip}
\keyword{algebra}
\keyword{array}
\keyword{methods}
