\name{not-methods}
\title{Logical Negation}
%
\docType{methods}
\keyword{logic}
\keyword{methods}
%
\alias{!}
\alias{!-methods}
%
\alias{!,Matrix-method}
\alias{!,indMatrix-method}
\alias{!,ldenseMatrix-method}
\alias{!,ldiMatrix-method}
\alias{!,lsparseMatrix-method}
\alias{!,lsparseVector-method}
\alias{!,ndenseMatrix-method}
\alias{!,ndiMatrix-method}
\alias{!,nsparseMatrix-method}
\alias{!,nsparseVector-method}
\alias{!,sparseVector-method}
%
\description{TODO.}
\usage{
\S4method{!}{ndenseMatrix}(x)
\S4method{!}{nsparseMatrix}(x)
\S4method{!}{ndiMatrix}(x)
\S4method{!}{nsparseVector}(x)

\S4method{!}{ldenseMatrix}(x)
\S4method{!}{lsparseMatrix}(x)
\S4method{!}{ldiMatrix}(x)
\S4method{!}{lsparseVector}(x)

\S4method{!}{Matrix}(x)
\S4method{!}{sparseVector}(x)
}
\arguments{
\item{x}{.}
}
%% \value{}
%% \seealso{}
%% \examples{}
