% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{op.log}
\alias{op.log}
\title{estimate parameters for the logisitic regression}
\usage{
op.log(Y, X, mod, alpha = 1)
}
\arguments{
\item{Y}{a vector with n 0-1 entry}

\item{X}{data}

\item{mod}{estimate mod}

\item{alpha}{alpha for the estimate}
}
\value{
estimated parameters, CI, log Likelihood
}
\description{
estimate parameters for the logisitic regression
}
