\name{Ops-methods}
\title{Arithmetic, Relational, and Logical Operators}
%
\docType{methods}
\keyword{arith}
\keyword{logic}
\keyword{methods}
%
\alias{Ops}
\alias{Ops-methods}
\alias{Arith}
\alias{Arith-methods}
\alias{Compare}
\alias{Compare-methods}
\alias{Logic}
\alias{Logic-methods}
%
\alias{Ops,ANY,iMatrix-method}
\alias{Ops,ANY,zMatrix-method}
\alias{Ops,Matrix,iMatrix-method}
\alias{Ops,Matrix,zMatrix-method}
\alias{Ops,iMatrix,ANY-method}
\alias{Ops,iMatrix,Matrix-method}
\alias{Ops,zMatrix,ANY-method}
\alias{Ops,zMatrix,Matrix-method}
%
\alias{Ops,ANY,Matrix-method}
\alias{Ops,ANY,ddiMatrix-method}
\alias{Ops,ANY,ldiMatrix-method}
\alias{Ops,ANY,sparseVector-method}
\alias{Ops,Matrix,ANY-method}
\alias{Ops,Matrix,NULL-method}
\alias{Ops,Matrix,ddiMatrix-method}
\alias{Ops,Matrix,ldiMatrix-method}
\alias{Ops,Matrix,matrix-method}
\alias{Ops,Matrix,sparseVector-method}
\alias{Ops,NULL,Matrix-method}
\alias{Ops,dMatrix,dMatrix-method}
\alias{Ops,dMatrix,ddiMatrix-method}
\alias{Ops,dMatrix,lMatrix-method}
\alias{Ops,dMatrix,ldiMatrix-method}
\alias{Ops,dMatrix,nMatrix-method}
\alias{Ops,ddiMatrix,ANY-method}
\alias{Ops,ddiMatrix,Matrix-method}
\alias{Ops,ddiMatrix,dMatrix-method}
\alias{Ops,ddiMatrix,ddiMatrix-method}
\alias{Ops,ddiMatrix,ldiMatrix-method}
\alias{Ops,ddiMatrix,logical-method}
\alias{Ops,ddiMatrix,ndiMatrix-method}
\alias{Ops,ddiMatrix,numeric-method}
\alias{Ops,ddiMatrix,sparseMatrix-method}
\alias{Ops,diagonalMatrix,triangularMatrix-method}
\alias{Ops,dpoMatrix,logical-method}
\alias{Ops,dpoMatrix,numeric-method}
\alias{Ops,dppMatrix,logical-method}
\alias{Ops,dppMatrix,numeric-method}
\alias{Ops,dsparseMatrix,nsparseMatrix-method}
\alias{Ops,lMatrix,dMatrix-method}
\alias{Ops,lMatrix,lMatrix-method}
\alias{Ops,lMatrix,nMatrix-method}
\alias{Ops,lMatrix,numeric-method}
\alias{Ops,ldenseMatrix,ldenseMatrix-method}
\alias{Ops,ldiMatrix,ANY-method}
\alias{Ops,ldiMatrix,Matrix-method}
\alias{Ops,ldiMatrix,dMatrix-method}
\alias{Ops,ldiMatrix,ddiMatrix-method}
\alias{Ops,ldiMatrix,ldiMatrix-method}
\alias{Ops,ldiMatrix,logical-method}
\alias{Ops,ldiMatrix,ndiMatrix-method}
\alias{Ops,ldiMatrix,numeric-method}
\alias{Ops,ldiMatrix,sparseMatrix-method}
\alias{Ops,logical,dpoMatrix-method}
\alias{Ops,logical,dppMatrix-method}
\alias{Ops,lsparseMatrix,lsparseMatrix-method}
\alias{Ops,lsparseMatrix,nsparseMatrix-method}
\alias{Ops,matrix,Matrix-method}
\alias{Ops,nMatrix,dMatrix-method}
\alias{Ops,nMatrix,lMatrix-method}
\alias{Ops,nMatrix,nMatrix-method}
\alias{Ops,nMatrix,numeric-method}
\alias{Ops,ndenseMatrix,ndenseMatrix-method}
\alias{Ops,ndiMatrix,ddiMatrix-method}
\alias{Ops,ndiMatrix,ldiMatrix-method}
\alias{Ops,ndiMatrix,ndiMatrix-method}
\alias{Ops,nsparseMatrix,dsparseMatrix-method}
\alias{Ops,nsparseMatrix,lsparseMatrix-method}
\alias{Ops,nsparseMatrix,sparseMatrix-method}
\alias{Ops,numeric,dpoMatrix-method}
\alias{Ops,numeric,dppMatrix-method}
\alias{Ops,numeric,lMatrix-method}
\alias{Ops,numeric,nMatrix-method}
\alias{Ops,numeric,sparseMatrix-method}
\alias{Ops,sparseMatrix,ddiMatrix-method}
\alias{Ops,sparseMatrix,ldiMatrix-method}
\alias{Ops,sparseMatrix,nsparseMatrix-method}
\alias{Ops,sparseMatrix,numeric-method}
\alias{Ops,sparseMatrix,sparseMatrix-method}
\alias{Ops,sparseVector,ANY-method}
\alias{Ops,sparseVector,Matrix-method}
\alias{Ops,sparseVector,sparseVector-method}
\alias{Ops,sparseVector,vector-method}
\alias{Ops,vector,sparseVector-method}
%
\alias{Arith,CsparseMatrix,CsparseMatrix-method}
\alias{Arith,CsparseMatrix,numeric-method}
\alias{Arith,Matrix,Matrix-method}
\alias{Arith,Matrix,lsparseMatrix-method}
\alias{Arith,Matrix,nsparseMatrix-method}
\alias{Arith,ddenseMatrix,logical-method}
\alias{Arith,ddenseMatrix,numeric-method}
\alias{Arith,ddenseMatrix,sparseVector-method}
\alias{Arith,ddiMatrix,logical-method}
\alias{Arith,ddiMatrix,numeric-method}
\alias{Arith,dgCMatrix,dgCMatrix-method}
\alias{Arith,dgCMatrix,logical-method}
\alias{Arith,dgCMatrix,numeric-method}
\alias{Arith,dgeMatrix,dgeMatrix-method}
\alias{Arith,dgeMatrix,logical-method}
\alias{Arith,dgeMatrix,numeric-method}
\alias{Arith,dgeMatrix,sparseVector-method}
\alias{Arith,dpoMatrix,logical-method}
\alias{Arith,dpoMatrix,numeric-method}
\alias{Arith,dppMatrix,logical-method}
\alias{Arith,dppMatrix,numeric-method}
\alias{Arith,dsCMatrix,dsCMatrix-method}
\alias{Arith,dsparseMatrix,logical-method}
\alias{Arith,dsparseMatrix,nsparseMatrix-method}
\alias{Arith,dsparseMatrix,numeric-method}
\alias{Arith,dsparseVector,dsparseVector-method}
\alias{Arith,dtCMatrix,dtCMatrix-method}
\alias{Arith,lMatrix,logical-method}
\alias{Arith,lMatrix,numeric-method}
\alias{Arith,ldiMatrix,logical-method}
\alias{Arith,ldiMatrix,numeric-method}
\alias{Arith,lgCMatrix,lgCMatrix-method}
\alias{Arith,lgTMatrix,lgTMatrix-method}
\alias{Arith,lgeMatrix,lgeMatrix-method}
\alias{Arith,logical,ddenseMatrix-method}
\alias{Arith,logical,ddiMatrix-method}
\alias{Arith,logical,dgCMatrix-method}
\alias{Arith,logical,dgeMatrix-method}
\alias{Arith,logical,dpoMatrix-method}
\alias{Arith,logical,dppMatrix-method}
\alias{Arith,logical,dsparseMatrix-method}
\alias{Arith,logical,lMatrix-method}
\alias{Arith,logical,ldiMatrix-method}
\alias{Arith,logical,nMatrix-method}
\alias{Arith,lsparseMatrix,Matrix-method}
\alias{Arith,lsparseMatrix,nsparseMatrix-method}
\alias{Arith,nMatrix,logical-method}
\alias{Arith,nMatrix,numeric-method}
\alias{Arith,ngeMatrix,ngeMatrix-method}
\alias{Arith,nsparseMatrix,Matrix-method}
\alias{Arith,nsparseMatrix,dsparseMatrix-method}
\alias{Arith,nsparseMatrix,lsparseMatrix-method}
\alias{Arith,numeric,CsparseMatrix-method}
\alias{Arith,numeric,ddenseMatrix-method}
\alias{Arith,numeric,ddiMatrix-method}
\alias{Arith,numeric,dgCMatrix-method}
\alias{Arith,numeric,dgeMatrix-method}
\alias{Arith,numeric,dpoMatrix-method}
\alias{Arith,numeric,dppMatrix-method}
\alias{Arith,numeric,dsparseMatrix-method}
\alias{Arith,numeric,lMatrix-method}
\alias{Arith,numeric,ldiMatrix-method}
\alias{Arith,numeric,nMatrix-method}
\alias{Arith,sparseVector,ddenseMatrix-method}
\alias{Arith,sparseVector,dgeMatrix-method}
\alias{Arith,sparseVector,sparseVector-method}
\alias{Arith,triangularMatrix,diagonalMatrix-method}
%
\alias{Compare,CsparseMatrix,CsparseMatrix-method}
\alias{Compare,dMatrix,logical-method}
\alias{Compare,dMatrix,numeric-method}
\alias{Compare,lMatrix,logical-method}
\alias{Compare,lMatrix,numeric-method}
\alias{Compare,lgeMatrix,lgeMatrix-method}
\alias{Compare,logical,dMatrix-method}
\alias{Compare,logical,lMatrix-method}
\alias{Compare,logical,nMatrix-method}
\alias{Compare,nMatrix,logical-method}
\alias{Compare,nMatrix,nMatrix-method}
\alias{Compare,nMatrix,numeric-method}
\alias{Compare,ngeMatrix,ngeMatrix-method}
\alias{Compare,numeric,dMatrix-method}
\alias{Compare,numeric,lMatrix-method}
\alias{Compare,numeric,nMatrix-method}
\alias{Compare,triangularMatrix,diagonalMatrix-method}
%
\alias{Logic,ANY,Matrix-method}
\alias{Logic,CsparseMatrix,CsparseMatrix-method}
\alias{Logic,Matrix,ANY-method}
\alias{Logic,Matrix,nMatrix-method}
\alias{Logic,dMatrix,logical-method}
\alias{Logic,dMatrix,numeric-method}
\alias{Logic,dMatrix,sparseVector-method}
\alias{Logic,lMatrix,logical-method}
\alias{Logic,lMatrix,numeric-method}
\alias{Logic,lMatrix,sparseVector-method}
\alias{Logic,ldenseMatrix,lsparseMatrix-method}
\alias{Logic,lgCMatrix,lgCMatrix-method}
\alias{Logic,lgTMatrix,lgTMatrix-method}
\alias{Logic,lgeMatrix,lgeMatrix-method}
\alias{Logic,logical,dMatrix-method}
\alias{Logic,logical,lMatrix-method}
\alias{Logic,logical,nMatrix-method}
\alias{Logic,lsCMatrix,lsCMatrix-method}
\alias{Logic,lsparseMatrix,ldenseMatrix-method}
\alias{Logic,lsparseMatrix,lsparseMatrix-method}
\alias{Logic,lsparseVector,lsparseVector-method}
\alias{Logic,ltCMatrix,ltCMatrix-method}
\alias{Logic,nMatrix,Matrix-method}
\alias{Logic,nMatrix,logical-method}
\alias{Logic,nMatrix,nMatrix-method}
\alias{Logic,nMatrix,numeric-method}
\alias{Logic,nMatrix,sparseVector-method}
\alias{Logic,ngeMatrix,ngeMatrix-method}
\alias{Logic,numeric,dMatrix-method}
\alias{Logic,numeric,lMatrix-method}
\alias{Logic,numeric,nMatrix-method}
\alias{Logic,sparseVector,dMatrix-method}
\alias{Logic,sparseVector,lMatrix-method}
\alias{Logic,sparseVector,nMatrix-method}
\alias{Logic,sparseVector,sparseVector-method}
\alias{Logic,triangularMatrix,diagonalMatrix-method}
%
\alias{&,Matrix,ddiMatrix-method}
\alias{&,Matrix,ldiMatrix-method}
\alias{&,Matrix,ndiMatrix-method}
\alias{&,ddenseMatrix,ddiMatrix-method}
\alias{&,ddenseMatrix,ldiMatrix-method}
\alias{&,ddenseMatrix,ndiMatrix-method}
\alias{&,ddiMatrix,Matrix-method}
\alias{&,ddiMatrix,ddenseMatrix-method}
\alias{&,ddiMatrix,ldenseMatrix-method}
\alias{&,ddiMatrix,ndenseMatrix-method}
\alias{&,ldenseMatrix,ddiMatrix-method}
\alias{&,ldenseMatrix,ldiMatrix-method}
\alias{&,ldenseMatrix,ndiMatrix-method}
\alias{&,ldiMatrix,Matrix-method}
\alias{&,ldiMatrix,ddenseMatrix-method}
\alias{&,ldiMatrix,ldenseMatrix-method}
\alias{&,ldiMatrix,ndenseMatrix-method}
\alias{&,ndenseMatrix,ddiMatrix-method}
\alias{&,ndenseMatrix,ldiMatrix-method}
\alias{&,ndenseMatrix,ndiMatrix-method}
\alias{&,ndiMatrix,Matrix-method}
\alias{&,ndiMatrix,ddenseMatrix-method}
\alias{&,ndiMatrix,ldenseMatrix-method}
\alias{&,ndiMatrix,ndenseMatrix-method}
%
\alias{*,Matrix,ddiMatrix-method}
\alias{*,Matrix,ldiMatrix-method}
\alias{*,Matrix,ndiMatrix-method}
\alias{*,ddenseMatrix,ddiMatrix-method}
\alias{*,ddenseMatrix,ldiMatrix-method}
\alias{*,ddenseMatrix,ndiMatrix-method}
\alias{*,ddiMatrix,Matrix-method}
\alias{*,ddiMatrix,ddenseMatrix-method}
\alias{*,ddiMatrix,ldenseMatrix-method}
\alias{*,ddiMatrix,ndenseMatrix-method}
\alias{*,ldenseMatrix,ddiMatrix-method}
\alias{*,ldenseMatrix,ldiMatrix-method}
\alias{*,ldenseMatrix,ndiMatrix-method}
\alias{*,ldiMatrix,Matrix-method}
\alias{*,ldiMatrix,ddenseMatrix-method}
\alias{*,ldiMatrix,ldenseMatrix-method}
\alias{*,ldiMatrix,ndenseMatrix-method}
\alias{*,ndenseMatrix,ddiMatrix-method}
\alias{*,ndenseMatrix,ldiMatrix-method}
\alias{*,ndenseMatrix,ndiMatrix-method}
\alias{*,ndiMatrix,Matrix-method}
\alias{*,ndiMatrix,Matrix-method}
\alias{*,ndiMatrix,ddenseMatrix-method}
\alias{*,ndiMatrix,ldenseMatrix-method}
\alias{*,ndiMatrix,ndenseMatrix-method}
%
\alias{+,Matrix,missing-method}
\alias{+,dgTMatrix,dgTMatrix-method}
%
\alias{-,Matrix,missing-method}
\alias{-,denseMatrix,missing-method}
\alias{-,diagonalMatrix,missing-method}
\alias{-,dsparseVector,missing-method}
\alias{-,indMatrix,missing-method}
\alias{-,nsparseMatrix,missing-method}
\alias{-,sparseMatrix,missing-method}
%
\alias{/,ddiMatrix,Matrix-method}
\alias{/,ddiMatrix,ddenseMatrix-method}
\alias{/,ddiMatrix,ldenseMatrix-method}
\alias{/,ddiMatrix,ndenseMatrix-method}
\alias{/,ldiMatrix,Matrix-method}
\alias{/,ldiMatrix,ddenseMatrix-method}
\alias{/,ldiMatrix,ldenseMatrix-method}
\alias{/,ldiMatrix,ndenseMatrix-method}
\alias{/,ndiMatrix,Matrix-method}
\alias{/,ndiMatrix,ddenseMatrix-method}
\alias{/,ndiMatrix,ldenseMatrix-method}
\alias{/,ndiMatrix,ndenseMatrix-method}
%
\alias{\%/\%,ddiMatrix,Matrix-method}
\alias{\%/\%,ddiMatrix,ddenseMatrix-method}
\alias{\%/\%,ddiMatrix,ldenseMatrix-method}
\alias{\%/\%,ddiMatrix,ndenseMatrix-method}
\alias{\%/\%,ldiMatrix,Matrix-method}
\alias{\%/\%,ldiMatrix,ddenseMatrix-method}
\alias{\%/\%,ldiMatrix,ldenseMatrix-method}
\alias{\%/\%,ldiMatrix,ndenseMatrix-method}
\alias{\%/\%,ndiMatrix,Matrix-method}
\alias{\%/\%,ndiMatrix,ddenseMatrix-method}
\alias{\%/\%,ndiMatrix,ldenseMatrix-method}
\alias{\%/\%,ndiMatrix,ndenseMatrix-method}
%
\alias{\%\%,ddiMatrix,Matrix-method}
\alias{\%\%,ddiMatrix,ddenseMatrix-method}
\alias{\%\%,ddiMatrix,ldenseMatrix-method}
\alias{\%\%,ddiMatrix,ndenseMatrix-method}
\alias{\%\%,ldiMatrix,Matrix-method}
\alias{\%\%,ldiMatrix,ddenseMatrix-method}
\alias{\%\%,ldiMatrix,ldenseMatrix-method}
\alias{\%\%,ldiMatrix,ndenseMatrix-method}
\alias{\%\%,ndiMatrix,Matrix-method}
\alias{\%\%,ndiMatrix,ddenseMatrix-method}
\alias{\%\%,ndiMatrix,ldenseMatrix-method}
\alias{\%\%,ndiMatrix,ndenseMatrix-method}
%
\alias{^,Matrix,ddiMatrix-method}
\alias{^,Matrix,ldiMatrix-method}
\alias{^,Matrix,ndiMatrix-method}
\alias{^,ddenseMatrix,ddiMatrix-method}
\alias{^,ddenseMatrix,ldiMatrix-method}
\alias{^,ddenseMatrix,ndiMatrix-method}
\alias{^,ldenseMatrix,ddiMatrix-method}
\alias{^,ldenseMatrix,ldiMatrix-method}
\alias{^,ldenseMatrix,ndiMatrix-method}
\alias{^,ndenseMatrix,ddiMatrix-method}
\alias{^,ndenseMatrix,ldiMatrix-method}
\alias{^,ndenseMatrix,ndiMatrix-method}
%
\description{TODO.}
\usage{
Ops(e1, e2)
Arith(e1, e2)
Compare(e1, e2)
Logic(e1, e2)
}
\arguments{
\item{e1}{.}
\item{e2}{.}
}
%% \value{}
%% \seealso{}
%% \examples{}

%% \item{Arith}{\code{signature(e1 = "dpoMatrix", e2 = "numeric")} (and
%%   quite a few other signatures): The result of (\dQuote{elementwise}
%%   defined) arithmetic operations is typically \emph{not}
%%   positive-definite anymore.  The only exceptions, currently, are
%%   multiplications, divisions or additions with \emph{positive}
%%   \code{length(.) == 1} numbers (or \code{\link{logical}}s).}
%% \item{"Math"}{\code{signature(x = "diagonalMatrix")}: all these
%%   group methods return a \code{"diagonalMatrix"}, apart from
%%   \code{\link{cumsum}()} etc which return a \emph{vector} also for
%%   \pkg{base} \code{\link{matrix}}.}
%% \item{*}{\code{signature(e1 = "ddiMatrix", e2="denseMatrix")}:
%%   arithmetic and other operators from the \code{\link[=S4groupGeneric]{Ops}}
%%   group have a few dozen explicit method definitions, in order to
%%   keep the results \emph{diagonal} in many cases, including the following:}
%% \item{/}{\code{signature(e1 = "ddiMatrix", e2="denseMatrix")}:
%%   the result is from class \code{\linkS4class{ddiMatrix}} which is
%%   typically very desirable.  Note that when \code{e2} contains
%%   off-diagonal zeros or \code{\link{NA}}s, we implicitly use \eqn{0 / x = 0}, hence
%%   differing from traditional \R arithmetic (where
%%   \eqn{0 / 0 \mapsto \mbox{NaN}}{0/0 |-> NaN}), in order to preserve
%%   sparsity.}
