% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.R
\name{QR_OLS}
\alias{QR_OLS}
\title{OLS by QR decomposition}
\usage{
QR_OLS(Y, X)
}
\arguments{
\item{Y}{a vector with n entry}

\item{X}{a n * p matrix}
}
\value{
estimated parameters for OLS regression by QR decomposition
}
\description{
OLS by QR decomposition
}
\examples{
Y=rnorm(100,10,1)
X=matrix(rnorm(10000,10,1),ncol=100)
QR_OLS(Y,X)
}
