\name{rep-methods}
\title{Repeat a Vector}
%
\docType{methods}
\keyword{manip}
\keyword{methods}
%
\alias{rep}
\alias{rep-methods}
%
\alias{rep,denseMatrix-method}
\alias{rep,sparseMatrix-method}
\alias{rep,sparseVector-method}
%
\description{TODO.}
\usage{
\S4method{rep}{sparseVector}(x, times, length.out, each, \dots)
}
\arguments{
\item{x}{.}
\item{times}{.}
\item{length.out}{.}
\item{each}{.}
\item{\dots}{.}
}
%% \value{}
%% \seealso{}
%% \examples{}
