\name{show-methods}
\title{Show a Matrix}
%
\docType{methods}
\keyword{character}
\keyword{print}
\keyword{methods}
%
\alias{show}
\alias{show-methods}
%
\alias{show,BunchKaufman-method}
\alias{show,Cholesky-method}
\alias{show,LU-method}
\alias{show,MatrixFactorization-method}
\alias{show,QR-method}
\alias{show,Schur-method}
\alias{show,denseMatrix-method}
\alias{show,diagonalMatrix-method}
\alias{show,sparseMatrix-method}
\alias{show,sparseVector-method}
%
\alias{print,diagonalMatrix-method}
\alias{print,sparseMatrix-method}
%
\alias{format,sparseMatrix-method}
%
\alias{summary,diagonalMatrix-method}
\alias{summary,sparseMatrix-method}
%
\alias{print.diagSummary}
\alias{print.sparseSummary}
%
\description{TODO.}
\usage{
show(object)
}
\arguments{
\item{object}{.}
}
%% \value{}
%% \seealso{}
%% \examples{}
