% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.R
\name{simu}
\alias{simu}
\title{generate MVN samples}
\usage{
simu(mu, sigma, N)
}
\arguments{
\item{mu}{a n dimensional numeric vector}

\item{sigma}{a n*n positive definite matrix}

\item{N}{a numeric value}
}
\value{
Sample data from MVN(mu,sigma) for N times
}
\description{
generate MVN samples
}
\examples{
sigma=matrix(c(1,.5,.5,.5,.5,1,5,.5,.5,.5,1,.5,.5,.5,.5,1),ncol=4)
data=simu(c(1,2,3,4), sigma,100)
summary(data)
cov(data)
}
