\name{Summary-methods}
\title{Summary Statistics}
%
\docType{methods}
\keyword{arith}
\keyword{logic}
\keyword{methods}
%
\alias{Summary}
\alias{Summary-methods}
%
\alias{Summary,CsparseMatrix-method}
\alias{Summary,RsparseMatrix-method}
\alias{Summary,TsparseMatrix-method}
\alias{Summary,denseMatrix-method}
\alias{Summary,diagonalMatrix-method}
\alias{Summary,indMatrix-method}
\alias{Summary,sparseVector-method}
%
\description{TODO.}
\usage{
Summary(x, \dots, na.rm = FALSE)
}
\arguments{
\item{x}{.}
\item{\dots}{.}
\item{na.rm}{.}
}
%% \value{}
%% \seealso{}
%% \examples{}
