\name{t-methods}
\title{Matrix Transpose}
%
\docType{methods}
\keyword{algebra}
\keyword{array}
\keyword{methods}
%
\alias{t}
\alias{t-methods}
%
\alias{t,CsparseMatrix-method}
\alias{t,RsparseMatrix-method}
\alias{t,TsparseMatrix-method}
\alias{t,denseMatrix-method}
\alias{t,diagonalMatrix-method}
\alias{t,indMatrix-method}
\alias{t,pMatrix-method}
\alias{t,sparseVector-method}
%
\description{TODO.}
\usage{
t(x)
}
\arguments{
\item{x}{.}
}
%% \value{}
%% \seealso{}
%% \examples{}
