\name{toeplitz-methods}
\title{Form a Toeplitz Matrix}
%
\docType{methods}
\keyword{array}
\keyword{ts}
\keyword{methods}
%
\alias{toeplitz}
\alias{toeplitz-methods}
%
\alias{toeplitz,sparseVector-method}
%
\description{TODO.}
\usage{
toeplitz(x, \dots)
\S4method{toeplitz}{sparseVector}(x, symmetric = TRUE, repr = c("C", "R", "T"),
         giveCsparse, \dots)
}
\arguments{
\item{x}{.}
\item{symmetric}{.}
\item{repr}{.}
\item{giveCsparse}{.}
\item{\dots}{.}
}
%% \value{}
%% \seealso{}
%% \examples{}
