%% This is just a stub, not yet used really: -- move to own page when used
\name{Unused-classes}
\title{Virtual Classes Not Yet Really Implemented and Used}
\docType{class}
\alias{iMatrix-class}
\alias{zMatrix-class}
\description{
  \code{iMatrix} is the virtual class of all \bold{i}nteger
  (S4) matrices.  It extends the \code{\linkS4class{Matrix}} class directly.

  \code{zMatrix} is the virtual class of all \code{\link{complex}}
  (S4) matrices.  It extends the \code{\linkS4class{Matrix}} class directly.
}
\examples{
showClass("iMatrix")
showClass("zMatrix")
}
\keyword{classes}

