\name{update-methods}
\title{Updating a Sparse Cholesky Factorization}
%
\docType{methods}
\keyword{algebra}
\keyword{array}
\keyword{methods}
%
\alias{update}
\alias{update-methods}
%
\alias{update,sparseCholesky-method}
%
\description{TODO.}
\usage{
update(object, \dots)
}
\arguments{
\item{object}{.}
\item{\dots}{.}
}
%% \value{}
%% \seealso{}
%% \examples{}
