\name{which-methods}
\title{Which indices are TRUE?}
%
\docType{methods}
\keyword{logic}
\keyword{methods}
%
\alias{which}
\alias{which-methods}
%
\alias{which,indMatrix-method}
\alias{which,ldenseMatrix-method}
\alias{which,ldiMatrix-method}
\alias{which,lsparseMatrix-method}
\alias{which,lsparseVector-method}
\alias{which,ndenseMatrix-method}
\alias{which,ndiMatrix-method}
\alias{which,nsparseMatrix-method}
\alias{which,nsparseVector-method}
%
\description{TODO.}
\usage{
which(x, arr.ind = FALSE, useNames = TRUE)
}
\arguments{
\item{x}{.}
\item{arr.ind}{.}
\item{useNames}{.}
}
%% \value{}
%% \seealso{}
%% \examples{}
