\name{y}
\alias{y}
\docType{data}
\title{A sample response vector}
\description{
The response vector of length 1850 corresponding to the model matrix \code{mm}.
}
\usage{data(y)}
\format{
  The format is:
 num [1:1850] 64.07  5.88 64.03  5.96 76.41 ...
}
%\details{}
%\source{}
\seealso{\code{\link{mm}}}
\references{
  Roger Koenker and Pin Ng (2003), ``SparseM: A sparse matrix package
  for R'', \emph{J. of Statistical Software}, 8(6).
}
\examples{
data(y)
}
\keyword{datasets}
