\name{zapsmall-methods}
\title{Zap Small Numbers to Zero}
%
\docType{methods}
\keyword{array}
\keyword{manip}
\keyword{methods}
%
\alias{zapsmall}
\alias{zapsmall-methods}
%
\alias{zapsmall,Matrix-method}
\alias{zapsmall,sparseVector-method}
%
\description{TODO.}

%% \item{zapsmall}{\code{signature(x = "Matrix")}: typically used for
%%   \code{"\linkS4class{dMatrix}"}: \code{\link{round}()} matrix entries
%%   such that (relatively) very small entries become zero exactly.}

%% \code{\link{drop0}(x, tol=1e-10)} is sometimes preferable to (and
%% more efficient than) \code{zapsmall(x, digits=10)}.
