\name{cov2cor-methods}
\title{Scale a Covariance Matrix}
%
\docType{methods}
\keyword{algebra}
\keyword{array}
\keyword{methods}
%
\alias{cov2cor}
\alias{cov2cor-methods}
%
\alias{cov2cor,denseMatrix-method}
\alias{cov2cor,sparseMatrix-method}
%
\description{TODO.}
\usage{
cov2cor(V)
}
\arguments{
\item{V}{.}
}
%% \value{}
%% \seealso{}
%% \examples{}
