\name{determinant-methods}
\title{Matrix Determinant}
%
\docType{methods}
\keyword{algebra}
\keyword{math}
\keyword{methods}
%
\alias{determinant}
\alias{determinant-methods}
%
\alias{determinant,Matrix,missing-method}
\alias{determinant,MatrixFactorization,missing-method}
\alias{determinant,denseBunchKaufman,logical-method}
\alias{determinant,denseCholesky,logical-method}
\alias{determinant,denseLU,logical-method}
\alias{determinant,denseSchur,logical-method}
\alias{determinant,diagonalMatrix,logical-method}
\alias{determinant,generalMatrix,logical-method}
\alias{determinant,indMatrix,logical-method}
\alias{determinant,pMatrix,logical-method}
\alias{determinant,sparseCholesky,logical-method}
\alias{determinant,sparseLU,logical-method}
\alias{determinant,sparseQR,logical-method}
\alias{determinant,symmetricMatrix,logical-method}
\alias{determinant,triangularMatrix,logical-method}
%
\alias{det}
%
\description{TODO.}
\usage{
det(x, \dots)
determinant(x, logarithm = TRUE, \dots)
}
\arguments{
\item{x}{.}
\item{logarithm}{.}
\item{\dots}{.}
}
%% \value{}
%% \seealso{}
%% \examples{}
