\name{dim-methods}
\title{Dimensions of a Matrix}
%
\docType{methods}
\keyword{array}
\keyword{attribute}
\keyword{methods}
%
\alias{dim}
\alias{dim-methods}
\alias{dim<-}
\alias{dim<--methods}
\alias{length}
\alias{length-methods}
\alias{length<-}
\alias{length<--methods}
%
\alias{dim,Matrix-method}
\alias{dim,MatrixFactorization-method}
\alias{dim,sparseVector-method}
%
\alias{dim<-,denseMatrix,numeric-method}
\alias{dim<-,sparseMatrix,numeric-method}
\alias{dim<-,sparseVector,numeric-method}
%
\alias{length,Matrix-method}
\alias{length,MatrixFactorization-method}
\alias{length,sparseVector-method}
%
\alias{length<-,denseMatrix,numeric-method}
\alias{length<-,sparseMatrix,numeric-method}
\alias{length<-,sparseVector,numeric-method}
%
\description{TODO.}
\usage{
\S4method{dim}{Matrix}(x)
\S4method{dim}{MatrixFactorization}(x)

\S4method{dim}{sparseVector}(x) <- value

\S4method{length}{Matrix}(x)
\S4method{length}{MatrixFactorization}(x)
\S4method{length}{sparseVector}(x)
}
\arguments{
\item{x}{.}
\item{value}{.}
}
%% \value{}
%% \seealso{}
%% \examples{}
