\name{Math-methods}
\title{Mathematical Functions}
%
\docType{methods}
\keyword{math}
\keyword{methods}
%
\alias{Math}
\alias{Math-methods}
\alias{log}
\alias{log-methods}
%
\alias{Math,denseMatrix-method}
\alias{Math,diagonalMatrix-method}
\alias{Math,indMatrix-method}
\alias{Math,sparseMatrix-method}
\alias{Math,sparseVector-method}
%
\alias{log,denseMatrix-method}
\alias{log,diagonalMatrix-method}
\alias{log,indMatrix-method}
\alias{log,sparseMatrix-method}
\alias{log,sparseVector-method}
%
\description{TODO.}
\usage{
Math(x)
## log(x, \dots)
}
\arguments{
\item{x}{.}
}
%% \value{}
%% \seealso{}
%% \examples{}
