\name{Matrix-class}
\docType{class}
\alias{Matrix-class}
\alias{show,Matrix-method}
\title{Class "Matrix" Virtual class of Matrices}
\description{The \code{Matrix} class is a virtual class contained by all actual
  classes in the \code{Matrix} package.}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
  \describe{
    \item{show}{\code{signature(object = "Matrix"}}
  }
}
%\references{}
\author{Douglas Bates \email{bates@stat.wisc.edu}}
%\note{}
\seealso{
  \code{\link{geMatrix-class}}, \code{\link{cscMatrix-class}}
}
%\examples{}
\keyword{classes}
\keyword{algebra}
