\name{Matrix-class}
\docType{class}
\alias{Matrix-class}
\alias{show,Matrix-method}
\title{Class "Matrix" Virtual Class of Matrices}
\description{
  The \code{Matrix} class is a virtual class contained by all actual
  classes in the \pkg{Matrix} package.
}
\section{Objects from the Class}{A virtual Class: No objects may be
  created from it.
}
\section{Methods}{
  \describe{
    \item{show}{\code{signature(object = "Matrix"}}
  }
}
%\references{}
\author{Douglas Bates \email{bates@stat.wisc.edu}}
\seealso{
  \code{\link{dgeMatrix-class}}, \code{\link{dgCMatrix-class}}, and
  \code{\link{Matrix}} for construction (and examples).
}
%\examples{}
\keyword{classes}
\keyword{algebra}
