\name{Matrix-notyet}
\title{Classes Not Yet Really Implemented and Used}
%
\docType{class}
\keyword{array}
\keyword{classes}
%
\alias{Matrix-notyet}
%
\alias{indexMatrix-class}
%
\description{
  \code{indexMatrix} is a virtual subclass of
  \code{\linkS4class{Matrix}} representing index matrices, defined here
  as matrices with exactly one nonzero entry per row or exactly one
  nonzero entry per column.
}
