\name{Ops-methods}
\title{Arithmetic, Relational, and Logical Operators}
%
\docType{methods}
\keyword{arith}
\keyword{logic}
\keyword{methods}
%
\alias{Ops}
\alias{Ops-methods}
\alias{Arith}
\alias{Arith-methods}
\alias{Compare}
\alias{Compare-methods}
\alias{Logic}
\alias{Logic-methods}
% ------------------------- existing S4 methods (most in ../R/Ops.R )
\alias{Ops,Matrix,NULL-method}
\alias{Ops,NULL,Matrix-method}
%
\alias{Ops,ANY,Matrix-method}
\alias{Ops,Matrix,ANY-method}
\alias{Ops,ANY,iMatrix-method}
\alias{Ops,ANY,zMatrix-method}
\alias{Ops,Matrix,iMatrix-method}
\alias{Ops,Matrix,zMatrix-method}
\alias{Ops,dMatrix,dMatrix-method}
\alias{Ops,dMatrix,lMatrix-method}
\alias{Ops,dMatrix,nMatrix-method}
\alias{Ops,iMatrix,ANY-method}
\alias{Ops,iMatrix,Matrix-method}
\alias{Ops,iMatrix,iMatrix-method}
\alias{Ops,lMatrix,dMatrix-method}
\alias{Ops,lMatrix,dMatrix-method}
\alias{Ops,lMatrix,lMatrix-method}
\alias{Ops,lMatrix,lMatrix-method}
\alias{Ops,lMatrix,nMatrix-method}
\alias{Ops,lMatrix,nMatrix-method}
\alias{Ops,nMatrix,dMatrix-method}
\alias{Ops,nMatrix,dMatrix-method}
\alias{Ops,nMatrix,lMatrix-method}
\alias{Ops,nMatrix,lMatrix-method}
\alias{Ops,nMatrix,nMatrix-method}
\alias{Ops,nMatrix,nMatrix-method}
\alias{Ops,zMatrix,ANY-method}
\alias{Ops,zMatrix,Matrix-method}
\alias{Ops,zMatrix,zMatrix-method}
%
\alias{Ops,ANY,ddiMatrix-method}
\alias{Ops,ANY,ldiMatrix-method}
\alias{Ops,ANY,sparseVector-method}
\alias{Ops,Matrix,ddiMatrix-method}
\alias{Ops,Matrix,ldiMatrix-method}
\alias{Ops,Matrix,matrix-method}
\alias{Ops,Matrix,sparseVector-method}
\alias{Ops,matrix,sparseVector-method}
\alias{Ops,sparseVector,matrix-method}
\alias{Ops,dMatrix,ddiMatrix-method}
\alias{Ops,dMatrix,ldiMatrix-method}
\alias{Ops,ddiMatrix,ANY-method}
\alias{Ops,ddiMatrix,Matrix-method}
\alias{Ops,ddiMatrix,dMatrix-method}
\alias{Ops,ddiMatrix,ddiMatrix-method}
\alias{Ops,ddiMatrix,ldiMatrix-method}
\alias{Ops,ddiMatrix,logical-method}
\alias{Ops,ddiMatrix,ndiMatrix-method}
\alias{Ops,ddiMatrix,numeric-method}
\alias{Ops,ddiMatrix,sparseMatrix-method}
\alias{Ops,diagonalMatrix,triangularMatrix-method}
\alias{Ops,ANY,diagonalMatrix-method}
\alias{Ops,Matrix,diagonalMatrix-method}
\alias{Ops,dMatrix,diagonalMatrix-method}
\alias{Ops,diagonalMatrix,ANY-method}
\alias{Ops,diagonalMatrix,Matrix-method}
\alias{Ops,diagonalMatrix,dMatrix-method}
\alias{Ops,diagonalMatrix,sparseMatrix-method}
\alias{Ops,sparseMatrix,diagonalMatrix-method}
\alias{Ops,dpoMatrix,logical-method}
\alias{Ops,dpoMatrix,numeric-method}
\alias{Ops,dppMatrix,logical-method}
\alias{Ops,dppMatrix,numeric-method}
\alias{Ops,dsparseMatrix,nsparseMatrix-method}
\alias{Ops,lMatrix,numeric-method}
\alias{Ops,ldenseMatrix,ldenseMatrix-method}
\alias{Ops,ldiMatrix,ANY-method}
\alias{Ops,ldiMatrix,Matrix-method}
\alias{Ops,ldiMatrix,dMatrix-method}
\alias{Ops,ldiMatrix,ddiMatrix-method}
\alias{Ops,ldiMatrix,ldiMatrix-method}
\alias{Ops,ldiMatrix,logical-method}
\alias{Ops,ldiMatrix,ndiMatrix-method}
\alias{Ops,ldiMatrix,numeric-method}
\alias{Ops,ldiMatrix,sparseMatrix-method}
\alias{Ops,logical,dpoMatrix-method}
\alias{Ops,logical,dppMatrix-method}
\alias{Ops,lsparseMatrix,lsparseMatrix-method}
\alias{Ops,lsparseMatrix,nsparseMatrix-method}
\alias{Ops,matrix,Matrix-method}
\alias{Ops,nMatrix,numeric-method}
\alias{Ops,ndenseMatrix,ndenseMatrix-method}
\alias{Ops,ndiMatrix,ddiMatrix-method}
\alias{Ops,ndiMatrix,ldiMatrix-method}
\alias{Ops,ndiMatrix,ndiMatrix-method}
\alias{Ops,nsparseMatrix,dsparseMatrix-method}
\alias{Ops,nsparseMatrix,lsparseMatrix-method}
\alias{Ops,nsparseMatrix,sparseMatrix-method}
\alias{Ops,numeric,dpoMatrix-method}
\alias{Ops,numeric,dppMatrix-method}
\alias{Ops,numeric,lMatrix-method}
\alias{Ops,numeric,nMatrix-method}
\alias{Ops,numeric,sparseMatrix-method}
\alias{Ops,sparseMatrix,ddiMatrix-method}
\alias{Ops,sparseMatrix,ldiMatrix-method}
\alias{Ops,sparseMatrix,nsparseMatrix-method}
\alias{Ops,sparseMatrix,numeric-method}
\alias{Ops,sparseMatrix,sparseMatrix-method}
\alias{Ops,sparseVector,ANY-method}
\alias{Ops,sparseVector,Matrix-method}
\alias{Ops,sparseVector,sparseVector-method}
\alias{Ops,sparseVector,vector-method}
\alias{Ops,vector,sparseVector-method}
\alias{Ops,ddiMatrix,idiMatrix-method}
\alias{Ops,ddiMatrix,zdiMatrix-method}
\alias{Ops,idiMatrix,ddiMatrix-method}
\alias{Ops,idiMatrix,idiMatrix-method}
\alias{Ops,idiMatrix,ldiMatrix-method}
\alias{Ops,idiMatrix,ndiMatrix-method}
\alias{Ops,idiMatrix,zdiMatrix-method}
\alias{Ops,ldiMatrix,idiMatrix-method}
\alias{Ops,ldiMatrix,zdiMatrix-method}
\alias{Ops,ndiMatrix,idiMatrix-method}
\alias{Ops,ndiMatrix,zdiMatrix-method}
\alias{Ops,zdiMatrix,ddiMatrix-method}
\alias{Ops,zdiMatrix,idiMatrix-method}
\alias{Ops,zdiMatrix,ldiMatrix-method}
\alias{Ops,zdiMatrix,ndiMatrix-method}
\alias{Ops,zdiMatrix,zdiMatrix-method}
%
\alias{Arith,complex,CsparseMatrix-method}
\alias{Arith,complex,dgCMatrix-method}
\alias{Arith,complex,dgeMatrix-method}
\alias{Arith,complex,diagonalMatrix-method}
\alias{Arith,complex,dsparseMatrix-method}
\alias{Arith,complex,igCMatrix-method}
\alias{Arith,complex,igeMatrix-method}
\alias{Arith,complex,isparseMatrix-method}
\alias{Arith,complex,lgCMatrix-method}
\alias{Arith,complex,lgeMatrix-method}
\alias{Arith,complex,lsparseMatrix-method}
\alias{Arith,complex,ngCMatrix-method}
\alias{Arith,complex,ngeMatrix-method}
\alias{Arith,complex,nsparseMatrix-method}
\alias{Arith,complex,zgCMatrix-method}
\alias{Arith,complex,zgeMatrix-method}
\alias{Arith,complex,zsparseMatrix-method}
\alias{Arith,CsparseMatrix,complex-method}
\alias{Arith,CsparseMatrix,CsparseMatrix-method}
\alias{Arith,CsparseMatrix,logical-method}
\alias{Arith,CsparseMatrix,numeric-method}
\alias{Arith,denseMatrix,logical-method}
\alias{Arith,denseMatrix,numeric-method}
\alias{Arith,denseMatrix,sparseVector-method}
\alias{Arith,dgCMatrix,complex-method}
\alias{Arith,dgCMatrix,dgCMatrix-method}
\alias{Arith,dgCMatrix,logical-method}
\alias{Arith,dgCMatrix,numeric-method}
\alias{Arith,dgeMatrix,complex-method}
\alias{Arith,dgeMatrix,dgeMatrix-method}
\alias{Arith,dgeMatrix,logical-method}
\alias{Arith,dgeMatrix,numeric-method}
\alias{Arith,dgeMatrix,sparseVector-method}
\alias{Arith,diagonalMatrix,complex-method}
\alias{Arith,diagonalMatrix,logical-method}
\alias{Arith,diagonalMatrix,numeric-method}
\alias{Arith,dpoMatrix,logical-method}
\alias{Arith,dpoMatrix,numeric-method}
\alias{Arith,dppMatrix,logical-method}
\alias{Arith,dppMatrix,numeric-method}
\alias{Arith,dsCMatrix,dsCMatrix-method}
\alias{Arith,dsparseMatrix,complex-method}
\alias{Arith,dsparseMatrix,logical-method}
\alias{Arith,dsparseMatrix,nsparseMatrix-method}
\alias{Arith,dsparseMatrix,numeric-method}
\alias{Arith,dsparseVector,dsparseVector-method}
\alias{Arith,dtCMatrix,dtCMatrix-method}
\alias{Arith,igCMatrix,complex-method}
\alias{Arith,igCMatrix,igCMatrix-method}
\alias{Arith,igCMatrix,logical-method}
\alias{Arith,igCMatrix,numeric-method}
\alias{Arith,igeMatrix,complex-method}
\alias{Arith,igeMatrix,igeMatrix-method}
\alias{Arith,igeMatrix,logical-method}
\alias{Arith,igeMatrix,numeric-method}
\alias{Arith,igeMatrix,sparseVector-method}
\alias{Arith,iMatrix,lMatrix-method}
\alias{Arith,iMatrix,nMatrix-method}
\alias{Arith,isCMatrix,isCMatrix-method}
\alias{Arith,isparseMatrix,complex-method}
\alias{Arith,isparseMatrix,logical-method}
\alias{Arith,isparseMatrix,numeric-method}
\alias{Arith,itCMatrix,itCMatrix-method}
\alias{Arith,lgCMatrix,complex-method}
\alias{Arith,lgCMatrix,lgCMatrix-method}
\alias{Arith,lgCMatrix,logical-method}
\alias{Arith,lgCMatrix,numeric-method}
\alias{Arith,lgeMatrix,complex-method}
\alias{Arith,lgeMatrix,lgeMatrix-method}
\alias{Arith,lgeMatrix,logical-method}
\alias{Arith,lgeMatrix,numeric-method}
\alias{Arith,lgeMatrix,sparseVector-method}
\alias{Arith,lgTMatrix,lgTMatrix-method}
\alias{Arith,lMatrix,iMatrix-method}
\alias{Arith,lMatrix,logical-method}
\alias{Arith,lMatrix,numeric-method}
\alias{Arith,logical,CsparseMatrix-method}
\alias{Arith,logical,denseMatrix-method}
\alias{Arith,logical,dgCMatrix-method}
\alias{Arith,logical,dgeMatrix-method}
\alias{Arith,logical,diagonalMatrix-method}
\alias{Arith,logical,dpoMatrix-method}
\alias{Arith,logical,dppMatrix-method}
\alias{Arith,logical,dsparseMatrix-method}
\alias{Arith,logical,igCMatrix-method}
\alias{Arith,logical,igeMatrix-method}
\alias{Arith,logical,isparseMatrix-method}
\alias{Arith,logical,lgCMatrix-method}
\alias{Arith,logical,lgeMatrix-method}
\alias{Arith,logical,lMatrix-method}
\alias{Arith,logical,lsparseMatrix-method}
\alias{Arith,logical,ngCMatrix-method}
\alias{Arith,logical,ngeMatrix-method}
\alias{Arith,logical,nMatrix-method}
\alias{Arith,logical,nsparseMatrix-method}
\alias{Arith,logical,zgCMatrix-method}
\alias{Arith,logical,zgeMatrix-method}
\alias{Arith,logical,zsparseMatrix-method}
\alias{Arith,lsCMatrix,lsCMatrix-method}
\alias{Arith,lsparseMatrix,complex-method}
\alias{Arith,lsparseMatrix,logical-method}
\alias{Arith,lsparseMatrix,Matrix-method}
\alias{Arith,lsparseMatrix,nsparseMatrix-method}
\alias{Arith,lsparseMatrix,numeric-method}
\alias{Arith,ltCMatrix,ltCMatrix-method}
\alias{Arith,Matrix,lsparseMatrix-method}
\alias{Arith,Matrix,Matrix-method}
\alias{Arith,Matrix,nsparseMatrix-method}
\alias{Arith,Matrix,zMatrix-method}
\alias{Arith,ngCMatrix,complex-method}
\alias{Arith,ngCMatrix,logical-method}
\alias{Arith,ngCMatrix,numeric-method}
\alias{Arith,ngeMatrix,complex-method}
\alias{Arith,ngeMatrix,logical-method}
\alias{Arith,ngeMatrix,ngeMatrix-method}
\alias{Arith,ngeMatrix,numeric-method}
\alias{Arith,ngeMatrix,sparseVector-method}
\alias{Arith,nMatrix,iMatrix-method}
\alias{Arith,nMatrix,logical-method}
\alias{Arith,nMatrix,numeric-method}
\alias{Arith,nsCMatrix,nsCMatrix-method}
\alias{Arith,nsparseMatrix,complex-method}
\alias{Arith,nsparseMatrix,dsparseMatrix-method}
\alias{Arith,nsparseMatrix,logical-method}
\alias{Arith,nsparseMatrix,lsparseMatrix-method}
\alias{Arith,nsparseMatrix,Matrix-method}
\alias{Arith,nsparseMatrix,numeric-method}
\alias{Arith,ntCMatrix,ntCMatrix-method}
\alias{Arith,numeric,CsparseMatrix-method}
\alias{Arith,numeric,denseMatrix-method}
\alias{Arith,numeric,dgCMatrix-method}
\alias{Arith,numeric,dgeMatrix-method}
\alias{Arith,numeric,diagonalMatrix-method}
\alias{Arith,numeric,dpoMatrix-method}
\alias{Arith,numeric,dppMatrix-method}
\alias{Arith,numeric,dsparseMatrix-method}
\alias{Arith,numeric,igCMatrix-method}
\alias{Arith,numeric,igeMatrix-method}
\alias{Arith,numeric,isparseMatrix-method}
\alias{Arith,numeric,lgCMatrix-method}
\alias{Arith,numeric,lgeMatrix-method}
\alias{Arith,numeric,lMatrix-method}
\alias{Arith,numeric,lsparseMatrix-method}
\alias{Arith,numeric,ngCMatrix-method}
\alias{Arith,numeric,ngeMatrix-method}
\alias{Arith,numeric,nMatrix-method}
\alias{Arith,numeric,nsparseMatrix-method}
\alias{Arith,numeric,zgCMatrix-method}
\alias{Arith,numeric,zgeMatrix-method}
\alias{Arith,numeric,zsparseMatrix-method}
\alias{Arith,sparseVector,denseMatrix-method}
\alias{Arith,sparseVector,dgeMatrix-method}
\alias{Arith,sparseVector,igeMatrix-method}
\alias{Arith,sparseVector,lgeMatrix-method}
\alias{Arith,sparseVector,ngeMatrix-method}
\alias{Arith,sparseVector,sparseVector-method}
\alias{Arith,sparseVector,zgeMatrix-method}
\alias{Arith,triangularMatrix,diagonalMatrix-method}
\alias{Arith,zgCMatrix,complex-method}
\alias{Arith,zgCMatrix,logical-method}
\alias{Arith,zgCMatrix,numeric-method}
\alias{Arith,zgCMatrix,zgCMatrix-method}
\alias{Arith,zgeMatrix,complex-method}
\alias{Arith,zgeMatrix,logical-method}
\alias{Arith,zgeMatrix,numeric-method}
\alias{Arith,zgeMatrix,sparseVector-method}
\alias{Arith,zgeMatrix,zgeMatrix-method}
\alias{Arith,zMatrix,Matrix-method}
\alias{Arith,zsCMatrix,zsCMatrix-method}
\alias{Arith,zsparseMatrix,complex-method}
\alias{Arith,zsparseMatrix,logical-method}
\alias{Arith,zsparseMatrix,numeric-method}
\alias{Arith,ztCMatrix,ztCMatrix-method}
%
\alias{Compare,CsparseMatrix,CsparseMatrix-method}
\alias{Compare,dMatrix,logical-method}
\alias{Compare,dMatrix,numeric-method}
\alias{Compare,lMatrix,logical-method}
\alias{Compare,lMatrix,numeric-method}
\alias{Compare,lgeMatrix,lgeMatrix-method}
\alias{Compare,logical,dMatrix-method}
\alias{Compare,logical,lMatrix-method}
\alias{Compare,logical,nMatrix-method}
\alias{Compare,nMatrix,logical-method}
\alias{Compare,nMatrix,nMatrix-method}
\alias{Compare,nMatrix,numeric-method}
\alias{Compare,ngeMatrix,ngeMatrix-method}
\alias{Compare,numeric,dMatrix-method}
\alias{Compare,numeric,lMatrix-method}
\alias{Compare,numeric,nMatrix-method}
\alias{Compare,triangularMatrix,diagonalMatrix-method}
\alias{Compare,complex,dMatrix-method}
\alias{Compare,complex,iMatrix-method}
\alias{Compare,complex,lMatrix-method}
\alias{Compare,complex,nMatrix-method}
\alias{Compare,complex,zMatrix-method}
\alias{Compare,dMatrix,complex-method}
\alias{Compare,iMatrix,complex-method}
\alias{Compare,iMatrix,logical-method}
\alias{Compare,iMatrix,numeric-method}
\alias{Compare,lMatrix,complex-method}
\alias{Compare,logical,iMatrix-method}
\alias{Compare,logical,zMatrix-method}
\alias{Compare,nMatrix,complex-method}
\alias{Compare,numeric,iMatrix-method}
\alias{Compare,numeric,zMatrix-method}
\alias{Compare,zMatrix,complex-method}
\alias{Compare,zMatrix,logical-method}
\alias{Compare,zMatrix,numeric-method}
%
\alias{Logic,ANY,Matrix-method}
\alias{Logic,CsparseMatrix,CsparseMatrix-method}
\alias{Logic,Matrix,ANY-method}
\alias{Logic,Matrix,nMatrix-method}
\alias{Logic,dMatrix,logical-method}
\alias{Logic,dMatrix,numeric-method}
\alias{Logic,dMatrix,sparseVector-method}
\alias{Logic,lMatrix,logical-method}
\alias{Logic,lMatrix,numeric-method}
\alias{Logic,lMatrix,sparseVector-method}
\alias{Logic,ldenseMatrix,lsparseMatrix-method}
\alias{Logic,lgCMatrix,lgCMatrix-method}
\alias{Logic,lgTMatrix,lgTMatrix-method}
\alias{Logic,lgeMatrix,lgeMatrix-method}
\alias{Logic,logical,dMatrix-method}
\alias{Logic,logical,lMatrix-method}
\alias{Logic,logical,nMatrix-method}
\alias{Logic,lsCMatrix,lsCMatrix-method}
\alias{Logic,lsparseMatrix,ldenseMatrix-method}
\alias{Logic,lsparseMatrix,lsparseMatrix-method}
\alias{Logic,lsparseVector,lsparseVector-method}
\alias{Logic,ltCMatrix,ltCMatrix-method}
\alias{Logic,nMatrix,Matrix-method}
\alias{Logic,nMatrix,logical-method}
\alias{Logic,nMatrix,nMatrix-method}
\alias{Logic,nMatrix,numeric-method}
\alias{Logic,nMatrix,sparseVector-method}
\alias{Logic,ngeMatrix,ngeMatrix-method}
\alias{Logic,numeric,dMatrix-method}
\alias{Logic,numeric,lMatrix-method}
\alias{Logic,numeric,nMatrix-method}
\alias{Logic,sparseVector,dMatrix-method}
\alias{Logic,sparseVector,lMatrix-method}
\alias{Logic,sparseVector,nMatrix-method}
\alias{Logic,sparseVector,sparseVector-method}
\alias{Logic,triangularMatrix,diagonalMatrix-method}
\alias{Logic,complex,dMatrix-method}
\alias{Logic,complex,iMatrix-method}
\alias{Logic,complex,lMatrix-method}
\alias{Logic,complex,nMatrix-method}
\alias{Logic,complex,zMatrix-method}
\alias{Logic,dMatrix,complex-method}
\alias{Logic,iMatrix,complex-method}
\alias{Logic,iMatrix,logical-method}
\alias{Logic,iMatrix,numeric-method}
\alias{Logic,iMatrix,sparseVector-method}
\alias{Logic,lMatrix,complex-method}
\alias{Logic,logical,iMatrix-method}
\alias{Logic,logical,zMatrix-method}
\alias{Logic,nMatrix,complex-method}
\alias{Logic,numeric,iMatrix-method}
\alias{Logic,numeric,zMatrix-method}
\alias{Logic,sparseVector,iMatrix-method}
\alias{Logic,sparseVector,zMatrix-method}
\alias{Logic,zMatrix,complex-method}
\alias{Logic,zMatrix,logical-method}
\alias{Logic,zMatrix,numeric-method}
\alias{Logic,zMatrix,sparseVector-method}
%
\alias{&,Matrix,ddiMatrix-method}
\alias{&,Matrix,idiMatrix-method}
\alias{&,Matrix,ldiMatrix-method}
\alias{&,Matrix,ndiMatrix-method}
\alias{&,Matrix,zdiMatrix-method}
\alias{&,ddenseMatrix,ddiMatrix-method}
\alias{&,ddenseMatrix,ldiMatrix-method}
\alias{&,ddenseMatrix,ndiMatrix-method}
\alias{&,ddiMatrix,Matrix-method}
\alias{&,ddiMatrix,ddenseMatrix-method}
\alias{&,ddiMatrix,ldenseMatrix-method}
\alias{&,ddiMatrix,ndenseMatrix-method}
\alias{&,ldenseMatrix,ddiMatrix-method}
\alias{&,ldenseMatrix,ldiMatrix-method}
\alias{&,ldenseMatrix,ndiMatrix-method}
\alias{&,ldiMatrix,Matrix-method}
\alias{&,ldiMatrix,ddenseMatrix-method}
\alias{&,ldiMatrix,ldenseMatrix-method}
\alias{&,ldiMatrix,ndenseMatrix-method}
\alias{&,ndenseMatrix,ddiMatrix-method}
\alias{&,ndenseMatrix,ldiMatrix-method}
\alias{&,ndenseMatrix,ndiMatrix-method}
\alias{&,ndiMatrix,Matrix-method}
\alias{&,ndiMatrix,ddenseMatrix-method}
\alias{&,ndiMatrix,ldenseMatrix-method}
\alias{&,ndiMatrix,ndenseMatrix-method}
\alias{&,ddenseMatrix,idiMatrix-method}
\alias{&,ddenseMatrix,zdiMatrix-method}
\alias{&,ddiMatrix,idenseMatrix-method}
\alias{&,ddiMatrix,zdenseMatrix-method}
\alias{&,idenseMatrix,ddiMatrix-method}
\alias{&,idenseMatrix,idiMatrix-method}
\alias{&,idenseMatrix,ldiMatrix-method}
\alias{&,idenseMatrix,ndiMatrix-method}
\alias{&,idenseMatrix,zdiMatrix-method}
\alias{&,idiMatrix,Matrix-method}
\alias{&,idiMatrix,ddenseMatrix-method}
\alias{&,idiMatrix,idenseMatrix-method}
\alias{&,idiMatrix,ldenseMatrix-method}
\alias{&,idiMatrix,ndenseMatrix-method}
\alias{&,idiMatrix,zdenseMatrix-method}
\alias{&,ldenseMatrix,idiMatrix-method}
\alias{&,ldenseMatrix,zdiMatrix-method}
\alias{&,ldiMatrix,idenseMatrix-method}
\alias{&,ldiMatrix,zdenseMatrix-method}
\alias{&,ndenseMatrix,idiMatrix-method}
\alias{&,ndenseMatrix,zdiMatrix-method}
\alias{&,ndiMatrix,idenseMatrix-method}
\alias{&,ndiMatrix,zdenseMatrix-method}
\alias{&,zdenseMatrix,ddiMatrix-method}
\alias{&,zdenseMatrix,idiMatrix-method}
\alias{&,zdenseMatrix,ldiMatrix-method}
\alias{&,zdenseMatrix,ndiMatrix-method}
\alias{&,zdenseMatrix,zdiMatrix-method}
\alias{&,zdiMatrix,Matrix-method}
\alias{&,zdiMatrix,ddenseMatrix-method}
\alias{&,zdiMatrix,idenseMatrix-method}
\alias{&,zdiMatrix,ldenseMatrix-method}
\alias{&,zdiMatrix,ndenseMatrix-method}
\alias{&,zdiMatrix,zdenseMatrix-method}
%
\alias{*,Matrix,ddiMatrix-method}
\alias{*,Matrix,ldiMatrix-method}
\alias{*,Matrix,ndiMatrix-method}
\alias{*,ddenseMatrix,ddiMatrix-method}
\alias{*,ddenseMatrix,ldiMatrix-method}
\alias{*,ddenseMatrix,ndiMatrix-method}
\alias{*,ddiMatrix,Matrix-method}
\alias{*,ddiMatrix,ddenseMatrix-method}
\alias{*,ddiMatrix,ldenseMatrix-method}
\alias{*,ddiMatrix,ndenseMatrix-method}
\alias{*,ldenseMatrix,ddiMatrix-method}
\alias{*,ldenseMatrix,ldiMatrix-method}
\alias{*,ldenseMatrix,ndiMatrix-method}
\alias{*,ldiMatrix,Matrix-method}
\alias{*,ldiMatrix,ddenseMatrix-method}
\alias{*,ldiMatrix,ldenseMatrix-method}
\alias{*,ldiMatrix,ndenseMatrix-method}
\alias{*,ndenseMatrix,ddiMatrix-method}
\alias{*,ndenseMatrix,ldiMatrix-method}
\alias{*,ndenseMatrix,ndiMatrix-method}
\alias{*,ndiMatrix,Matrix-method}
\alias{*,ndiMatrix,Matrix-method}
\alias{*,ndiMatrix,ddenseMatrix-method}
\alias{*,ndiMatrix,ldenseMatrix-method}
\alias{*,ndiMatrix,ndenseMatrix-method}
\alias{*,Matrix,idiMatrix-method}
\alias{*,Matrix,zdiMatrix-method}
\alias{*,ddenseMatrix,idiMatrix-method}
\alias{*,ddenseMatrix,zdiMatrix-method}
\alias{*,ddiMatrix,idenseMatrix-method}
\alias{*,ddiMatrix,zdenseMatrix-method}
\alias{*,idenseMatrix,ddiMatrix-method}
\alias{*,idenseMatrix,idiMatrix-method}
\alias{*,idenseMatrix,ldiMatrix-method}
\alias{*,idenseMatrix,ndiMatrix-method}
\alias{*,idenseMatrix,zdiMatrix-method}
\alias{*,idiMatrix,Matrix-method}
\alias{*,idiMatrix,ddenseMatrix-method}
\alias{*,idiMatrix,idenseMatrix-method}
\alias{*,idiMatrix,ldenseMatrix-method}
\alias{*,idiMatrix,ndenseMatrix-method}
\alias{*,idiMatrix,zdenseMatrix-method}
\alias{*,ldenseMatrix,idiMatrix-method}
\alias{*,ldenseMatrix,zdiMatrix-method}
\alias{*,ldiMatrix,idenseMatrix-method}
\alias{*,ldiMatrix,zdenseMatrix-method}
\alias{*,ndenseMatrix,idiMatrix-method}
\alias{*,ndenseMatrix,zdiMatrix-method}
\alias{*,ndiMatrix,idenseMatrix-method}
\alias{*,ndiMatrix,zdenseMatrix-method}
\alias{*,zdenseMatrix,ddiMatrix-method}
\alias{*,zdenseMatrix,idiMatrix-method}
\alias{*,zdenseMatrix,ldiMatrix-method}
\alias{*,zdenseMatrix,ndiMatrix-method}
\alias{*,zdenseMatrix,zdiMatrix-method}
\alias{*,zdiMatrix,Matrix-method}
\alias{*,zdiMatrix,ddenseMatrix-method}
\alias{*,zdiMatrix,idenseMatrix-method}
\alias{*,zdiMatrix,ldenseMatrix-method}
\alias{*,zdiMatrix,ndenseMatrix-method}
\alias{*,zdiMatrix,zdenseMatrix-method}
%
\alias{+,Matrix,missing-method}
\alias{+,dgTMatrix,dgTMatrix-method}
\alias{+,sparseVector,missing-method}
%
\alias{-,Matrix,missing-method}
\alias{-,sparseVector,missing-method}
%
\alias{/,ddiMatrix,Matrix-method}
\alias{/,ddiMatrix,ddenseMatrix-method}
\alias{/,ddiMatrix,ldenseMatrix-method}
\alias{/,ddiMatrix,ndenseMatrix-method}
\alias{/,ldiMatrix,Matrix-method}
\alias{/,ldiMatrix,ddenseMatrix-method}
\alias{/,ldiMatrix,ldenseMatrix-method}
\alias{/,ldiMatrix,ndenseMatrix-method}
\alias{/,ndiMatrix,Matrix-method}
\alias{/,ndiMatrix,ddenseMatrix-method}
\alias{/,ndiMatrix,ldenseMatrix-method}
\alias{/,ndiMatrix,ndenseMatrix-method}
\alias{/,ddiMatrix,idenseMatrix-method}
\alias{/,ddiMatrix,zdenseMatrix-method}
\alias{/,idiMatrix,Matrix-method}
\alias{/,idiMatrix,ddenseMatrix-method}
\alias{/,idiMatrix,idenseMatrix-method}
\alias{/,idiMatrix,ldenseMatrix-method}
\alias{/,idiMatrix,ndenseMatrix-method}
\alias{/,idiMatrix,zdenseMatrix-method}
\alias{/,ldiMatrix,idenseMatrix-method}
\alias{/,ldiMatrix,zdenseMatrix-method}
\alias{/,ndiMatrix,idenseMatrix-method}
\alias{/,ndiMatrix,zdenseMatrix-method}
\alias{/,zdiMatrix,Matrix-method}
\alias{/,zdiMatrix,ddenseMatrix-method}
\alias{/,zdiMatrix,idenseMatrix-method}
\alias{/,zdiMatrix,ldenseMatrix-method}
\alias{/,zdiMatrix,ndenseMatrix-method}
\alias{/,zdiMatrix,zdenseMatrix-method}
%
\alias{\%/\%,ddiMatrix,Matrix-method}
\alias{\%/\%,ddiMatrix,ddenseMatrix-method}
\alias{\%/\%,ddiMatrix,ldenseMatrix-method}
\alias{\%/\%,ddiMatrix,ndenseMatrix-method}
\alias{\%/\%,ldiMatrix,Matrix-method}
\alias{\%/\%,ldiMatrix,ddenseMatrix-method}
\alias{\%/\%,ldiMatrix,ldenseMatrix-method}
\alias{\%/\%,ldiMatrix,ndenseMatrix-method}
\alias{\%/\%,ndiMatrix,Matrix-method}
\alias{\%/\%,ndiMatrix,ddenseMatrix-method}
\alias{\%/\%,ndiMatrix,ldenseMatrix-method}
\alias{\%/\%,ndiMatrix,ndenseMatrix-method}
\alias{\%/\%,ddiMatrix,idenseMatrix-method}
\alias{\%/\%,ddiMatrix,zdenseMatrix-method}
\alias{\%/\%,idiMatrix,Matrix-method}
\alias{\%/\%,idiMatrix,ddenseMatrix-method}
\alias{\%/\%,idiMatrix,idenseMatrix-method}
\alias{\%/\%,idiMatrix,ldenseMatrix-method}
\alias{\%/\%,idiMatrix,ndenseMatrix-method}
\alias{\%/\%,idiMatrix,zdenseMatrix-method}
\alias{\%/\%,ldiMatrix,idenseMatrix-method}
\alias{\%/\%,ldiMatrix,zdenseMatrix-method}
\alias{\%/\%,ndiMatrix,idenseMatrix-method}
\alias{\%/\%,ndiMatrix,zdenseMatrix-method}
\alias{\%/\%,zdiMatrix,Matrix-method}
\alias{\%/\%,zdiMatrix,ddenseMatrix-method}
\alias{\%/\%,zdiMatrix,idenseMatrix-method}
\alias{\%/\%,zdiMatrix,ldenseMatrix-method}
\alias{\%/\%,zdiMatrix,ndenseMatrix-method}
\alias{\%/\%,zdiMatrix,zdenseMatrix-method}
%
\alias{\%\%,ddiMatrix,Matrix-method}
\alias{\%\%,ddiMatrix,ddenseMatrix-method}
\alias{\%\%,ddiMatrix,ldenseMatrix-method}
\alias{\%\%,ddiMatrix,ndenseMatrix-method}
\alias{\%\%,ldiMatrix,Matrix-method}
\alias{\%\%,ldiMatrix,ddenseMatrix-method}
\alias{\%\%,ldiMatrix,ldenseMatrix-method}
\alias{\%\%,ldiMatrix,ndenseMatrix-method}
\alias{\%\%,ndiMatrix,Matrix-method}
\alias{\%\%,ndiMatrix,ddenseMatrix-method}
\alias{\%\%,ndiMatrix,ldenseMatrix-method}
\alias{\%\%,ndiMatrix,ndenseMatrix-method}
\alias{\%\%,ddiMatrix,idenseMatrix-method}
\alias{\%\%,ddiMatrix,zdenseMatrix-method}
\alias{\%\%,idiMatrix,Matrix-method}
\alias{\%\%,idiMatrix,ddenseMatrix-method}
\alias{\%\%,idiMatrix,idenseMatrix-method}
\alias{\%\%,idiMatrix,ldenseMatrix-method}
\alias{\%\%,idiMatrix,ndenseMatrix-method}
\alias{\%\%,idiMatrix,zdenseMatrix-method}
\alias{\%\%,ldiMatrix,idenseMatrix-method}
\alias{\%\%,ldiMatrix,zdenseMatrix-method}
\alias{\%\%,ndiMatrix,idenseMatrix-method}
\alias{\%\%,ndiMatrix,zdenseMatrix-method}
\alias{\%\%,zdiMatrix,Matrix-method}
\alias{\%\%,zdiMatrix,ddenseMatrix-method}
\alias{\%\%,zdiMatrix,idenseMatrix-method}
\alias{\%\%,zdiMatrix,ldenseMatrix-method}
\alias{\%\%,zdiMatrix,ndenseMatrix-method}
\alias{\%\%,zdiMatrix,zdenseMatrix-method}
%
\alias{^,Matrix,ddiMatrix-method}
\alias{^,Matrix,ldiMatrix-method}
\alias{^,Matrix,ndiMatrix-method}
\alias{^,Matrix,idiMatrix-method}
\alias{^,Matrix,zdiMatrix-method}
\alias{^,ddenseMatrix,ddiMatrix-method}
\alias{^,ddenseMatrix,ldiMatrix-method}
\alias{^,ddenseMatrix,ndiMatrix-method}
\alias{^,ldenseMatrix,ddiMatrix-method}
\alias{^,ldenseMatrix,ldiMatrix-method}
\alias{^,ldenseMatrix,ndiMatrix-method}
\alias{^,ndenseMatrix,ddiMatrix-method}
\alias{^,ndenseMatrix,ldiMatrix-method}
\alias{^,ndenseMatrix,ndiMatrix-method}
\alias{^,ddenseMatrix,idiMatrix-method}
\alias{^,ddenseMatrix,zdiMatrix-method}
\alias{^,idenseMatrix,ddiMatrix-method}
\alias{^,idenseMatrix,idiMatrix-method}
\alias{^,idenseMatrix,ldiMatrix-method}
\alias{^,idenseMatrix,ndiMatrix-method}
\alias{^,idenseMatrix,zdiMatrix-method}
\alias{^,ldenseMatrix,idiMatrix-method}
\alias{^,ldenseMatrix,zdiMatrix-method}
\alias{^,ndenseMatrix,idiMatrix-method}
\alias{^,ndenseMatrix,zdiMatrix-method}
\alias{^,zdenseMatrix,ddiMatrix-method}
\alias{^,zdenseMatrix,idiMatrix-method}
\alias{^,zdenseMatrix,ldiMatrix-method}
\alias{^,zdenseMatrix,ndiMatrix-method}
\alias{^,zdenseMatrix,zdiMatrix-method}
%
\description{
  Many \pkg{Matrix} package (binary operator) methods for \pkg{Matrix}
  objects (of super classes \code{\linkS4class{Matrix}} and
  \code{\linkS4class{sparseVector}}) combined with (base)
  \code{\link{matrix}}, and atomic vectors \code{\link{numeric}}, \code{\link{logical}},
  \code{\link{complex}}, etc.

  Methods are both defined as group methods, e.g., for the \code{"Arith"}
  group, and as regular methods for specific binary operators, e.g., \code{+}.
}
\usage{
Ops(e1, e2)
Arith(e1, e2)
Compare(e1, e2)
Logic(e1, e2)

#  e1 + e2
#  e1 / e2
#  e1 | e2    # etc
}
\arguments{
  \item{e1,e2}{typically \dQuote{numeric-alike} vectors,
    \code{"\linkS4class{sparseVector}"}, or \code{"\linkS4class{Matrix}"}
    objects.}
}
\value{typically a class \code{"\linkS4class{Matrix}"} object of a class
  related to the classes of \code{e1} and \code{e2}.
}
\seealso{Package \pkg{methods}: The overview \code{\link{Methods}};
  further \code{\link{getGroup}}, \code{\link{getGroupMembers}}
}
\examples{
## 'Ops': the union group of the 3 binary op groups:
getGroupMembers("Ops") #  "Arith"  "Compare" "Logic"
##
getGroupMembers("Arith")   # "+"  "-"  "*"  "^"  "%%"  "%/%"  "/"
getGroupMembers("Compare") # "==" ">"  "<"  "!=" "<="  ">="
getGroupMembers("Logic")   # "&"  "|"           ( *not* containing  unary "!" )
}
%% \value{}
%% \seealso{}
%% \examples{}

%% \item{Arith}{\code{signature(e1 = "dpoMatrix", e2 = "numeric")} (and
%%   quite a few other signatures): The result of (\dQuote{elementwise}
%%   defined) arithmetic operations is typically \emph{not}
%%   positive-definite anymore.  The only exceptions, currently, are
%%   multiplications, divisions or additions with \emph{positive}
%%   \code{length(.) == 1} numbers (or \code{\link{logical}}s).}
%% \item{"Math"}{\code{signature(x = "diagonalMatrix")}: all these
%%   group methods return a \code{"diagonalMatrix"}, apart from
%%   \code{\link{cumsum}()} etc which return a \emph{vector} also for
%%   \pkg{base} \code{\link{matrix}}.}
%% \item{*}{\code{signature(e1 = "ddiMatrix", e2="denseMatrix")}:
%%   arithmetic and other operators from the \code{\link[=S4groupGeneric]{Ops}}
%%   group have a few dozen explicit method definitions, in order to
%%   keep the results \emph{diagonal} in many cases, including the following:}
%% \item{/}{\code{signature(e1 = "ddiMatrix", e2="denseMatrix")}:
%%   the result is from class \code{\linkS4class{ddiMatrix}} which is
%%   typically very desirable.  Note that when \code{e2} contains
%%   off-diagonal zeros or \code{\link{NA}}s, we implicitly use \eqn{0 / x = 0}, hence
%%   differing from traditional \R arithmetic (where
%%   \eqn{0 / 0 \mapsto \mbox{NaN}}{0/0 |-> NaN}), in order to preserve
%%   sparsity.}
