\name{sort-methods}
\title{Sort a Vector}
%
\docType{methods}
\keyword{manip}
\keyword{methods}
%
\alias{sort}
\alias{sort-methods}
%
\alias{sort,denseMatrix-method}
\alias{sort,sparseMatrix-method}
\alias{sort,sparseVector-method}
%
\description{TODO.}
\usage{
\S4method{sort}{sparseVector}(x, decreasing = FALSE, na.last = NA, \dots)
}
\arguments{
\item{x}{.}
\item{decreasing}{.}
\item{na.last}{.}
\item{\dots}{.}
}
%% \value{}
%% \seealso{}
%% \examples{}
