\name{t-methods}
\title{Matrix (Conjugate) Transpose}
%
\docType{methods}
\keyword{algebra}
\keyword{array}
\keyword{methods}
%
\alias{t}
\alias{t-methods}
\alias{ct}
\alias{ct-methods}
%
\alias{t,CsparseMatrix-method}
\alias{t,RsparseMatrix-method}
\alias{t,TsparseMatrix-method}
\alias{t,denseMatrix-method}
\alias{t,diagonalMatrix-method}
\alias{t,indMatrix-method}
\alias{t,pMatrix-method}
\alias{t,sparseVector-method}
%
\alias{ct,CsparseMatrix-method}
\alias{ct,RsparseMatrix-method}
\alias{ct,TsparseMatrix-method}
\alias{ct,denseMatrix-method}
\alias{ct,diagonalMatrix-method}
\alias{ct,indMatrix-method}
\alias{ct,matrix-method}
\alias{ct,pMatrix-method}
\alias{ct,sparseVector-method}
%
\description{TODO.}
\usage{
 t(x)
ct(x)
}
\arguments{
\item{x}{a matrix or \code{\linkS4class{Matrix}}.}
}
%% \value{}
%% \seealso{}
%% \examples{}

%% \item{\code{t}}{\code{signature(x = "indMatrix")}:
%%   the transpose, which is an \code{indMatrix} with identical
%%   \code{perm} but opposite \code{margin}.}
