\name{which-methods}
\title{Which indices are TRUE?}
%
\docType{methods}
\keyword{logic}
\keyword{methods}
%
\alias{which}
\alias{which-methods}
%
\alias{which,indMatrix-method}
\alias{which,ldenseMatrix-method}
\alias{which,ldiMatrix-method}
\alias{which,lsparseMatrix-method}
\alias{which,lsparseVector-method}
\alias{which,ndenseMatrix-method}
\alias{which,ndiMatrix-method}
\alias{which,nsparseMatrix-method}
\alias{which,nsparseVector-method}
%
\description{TODO.}
\usage{
which(x, arr.ind = FALSE, useNames = TRUE)
}
\arguments{
\item{x}{.}
\item{arr.ind}{.}
\item{useNames}{.}
}
%% \value{}
%% \seealso{}
%% \examples{}

%% \item{which}{\code{signature(x = "lsparseMatrix")} and many other
%%   subclasses of \code{"lMatrix"}: as the \pkg{base} function
%%   \code{\link{which}(x, arr.ind)} returns the indices of the
%%   \code{\link{TRUE}} entries in \code{x}; if \code{arr.ind} is true,
%%   as a 2-column matrix of row and column indices. Since \pkg{Matrix}
%%   version 1.2-9, if \code{useNames} is true, as by default, with
%%   \code{\link{dimnames}}, the same as \code{base::which}.}
